pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "titanflow"
version = "0.1.0"
description = "TitanFlow — Orchestration engine for TitanArray"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "python-telegram-bot>=21.0",
    "apscheduler>=3.10.0",
    "sqlmodel>=0.0.22",
    "aiosqlite>=0.20.0",
    "pydantic-settings>=2.5.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "feedparser>=6.0.0",
    "ollama>=0.4.0",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"
